BookingForm: Refactor Fields
