+++ lib/src/features/bookings/presentation/screens/edit_booking_screen/widgets/form_fields/date_picker_field.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:keeley/src/constants/keys.dart';
 import 'package:keeley/src/constants/strings.dart';
-import 'package:keeley/src/features/bookings/presentation/controllers/bookings_controller.dart';
+import 'package:keeley/src/features/bookings/presentation/controllers/edit_booking_controller.dart';
 import 'package:shadcn_ui/shadcn_ui.dart';
 
 class DatePickerField extends ConsumerWidget {
@@ -30,7 +30,7 @@ class DatePickerField extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final bookingState = ref.watch(bookingsControllerProvider);
+    final bookingState = ref.watch(editBookingControllerProvider);
 
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
